use lean-com

+++ packs/Nuvoton/NuMicroM55_DFP/3.1.4-rc.3/Library/StdDriver/src/retarget.c
@@ -217,7 +217,7 @@ static char GetChar(void)
 
     if (g_ICE_Connected)
     {
-        int nRet;
+        long int nRet;
 
 #if defined (__ICCARM__)
